Refactored autosave services file names
