+++ server/planning/planning/module.py
@@ -15,7 +15,7 @@ from planning.types import (
 from .planning_service import PlanningAsyncService
 from .planning_history_async_service import PlanningHistoryAsyncService
 from .planning_featured_async_service import PlanningFeaturedAsyncService
-from .planning_autosave_async_service import PlanningAutosaveAsyncService
+from .planning_autosave_service import PlanningAutosaveAsyncService
 
 planning_resource_config: ResourceConfig = ResourceConfig(
     name="planning",
