+++ server/planning/planning/planning.py
@@ -46,7 +46,7 @@ from planning.types import (
     PLANNING_RELATED_EVENT_LINK_TYPE,
 )
 from planning.planning.planning_history_async_service import PlanningHistoryAsyncService
-from planning.planning.planning_autosave_async_service import PlanningAutosaveAsyncService
+from planning.planning.planning_autosave_service import PlanningAutosaveAsyncService
 from planning.assignments.assignments_history_async import AssignmentsHistoryAsyncService
 from planning.common import (
     get_coverage_status_from_cv,
